# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Leave Management API
  description: "Leave requests with a Manager → HR approval chain, escalation, delegation\
    \ and conflict flags."
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/notifications/{id}/read:
    post:
      tags:
      - Notifications
      summary: Mark one of the caller's notifications as read
      operationId: markRead
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotificationDto"
      security:
      - bearerAuth: []
...
  /api/notifications:
    get:
      tags:
      - Notifications
      summary: "The caller's in-app notifications, newest first"
      operationId: list_2
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotificationDto"
      security:
      - bearerAuth: []
  /api/manager/requests:
    get:
      tags:
      - Manager
      summary: "Requests in the caller's approval scope (own team plus delegated),\
        \ filterable and paged"
      operationId: requests
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - PENDING_MANAGER
          - ESCALATED
          - PENDING_HR
          - APPROVED
          - REJECTED
          - CANCELLED
      - name: type
        in: query
        required: false
        schema:
          type: string
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: sort
        in: query
        required: false
        schema:
          type: string
          default: "createdAt,desc"
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseLeaveRequestDto"
      security:
      - bearerAuth: []
...
  /api/hr/requests:
    get:
      tags:
      - HR
      summary: "All requests organisation-wide, filterable and paged"
      operationId: requests_1
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - PENDING_MANAGER
          - ESCALATED
          - PENDING_HR
          - APPROVED
          - REJECTED
          - CANCELLED
      - name: type
        in: query
        required: false
        schema:
          type: string
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: sort
        in: query
        required: false
        schema:
          type: string
          default: "createdAt,desc"
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseLeaveRequestDto"
      security:
      - bearerAuth: []
...
  /api/analytics/summary:
    get:
      tags:
      - Analytics
      summary: Organisation-wide leave analytics summary
      operationId: summary
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AnalyticsSummaryDto"
      security:
      - bearerAuth: []
...
components:
  schemas:
    NotificationDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        message:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
        relatedRequestId:
          type: integer
          format: int64
...
    ApprovalStepDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        stage:
          type: string
          enum:
          - MANAGER
          - HR
        assigneeId:
          type: integer
          format: int64
        assigneeName:
          type: string
        delegatedFromName:
          type: string
        dueAt:
          type: string
          format: date-time
        decision:
          type: string
          enum:
          - PENDING
          - APPROVED
          - REJECTED
          - ESCALATED
        decidedAt:
          type: string
          format: date-time
        comment:
          type: string
    LeaveRequestDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        employeeId:
          type: integer
          format: int64
        employeeName:
          type: string
        teamName:
          type: string
        leaveTypeCode:
          type: string
        leaveTypeName:
          type: string
        fromDate:
          type: string
          format: date
        toDate:
          type: string
          format: date
        days:
          type: number
        reason:
          type: string
        status:
          type: string
          enum:
          - PENDING_MANAGER
          - ESCALATED
          - PENDING_HR
          - APPROVED
          - REJECTED
          - CANCELLED
        flagged:
          type: boolean
        flagReason:
          type: string
        createdAt:
          type: string
          format: date-time
        lastActionAt:
          type: string
          format: date-time
        steps:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalStepDto"
        canApprove:
          type: boolean
        canCancel:
          type: boolean
...
    PageResponseLeaveRequestDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/LeaveRequestDto"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    AnalyticsSummaryDto:
      type: object
      properties:
        totalRequests:
          type: integer
          format: int64
        byStatus:
          type: object
          additionalProperties:
            type: integer
            format: int64
        byLeaveType:
          type: object
          additionalProperties:
            type: integer
            format: int64
        flaggedCount:
          type: integer
          format: int64
        escalatedCount:
          type: integer
          format: int64
        avgDecisionHours:
          type: number
          format: double
        monthlyTrend:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyPoint"
        teamLoad:
          type: array
          items:
            $ref: "#/components/schemas/TeamLoad"
    MonthlyPoint:
      type: object
      properties:
        month:
          type: string
        days:
          type: number
          format: double
        requests:
          type: integer
          format: int64
    TeamLoad:
      type: object
      properties:
        teamName:
          type: string
        pending:
          type: integer
          format: int64
        approved:
          type: integer
          format: int64
        flagged:
          type: integer
          format: int64
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT